\name{RcmdrcoinPlugin-internal}
\title{Internal RcmdrcoinPlugin objects}
\alias{fncCoinTwoSampleWilcoxonTest}
\alias{fncCoinPairedWilcoxonTest}
\alias{fncCoinMedianTest}
\alias{fncCoinNormalTest}
\alias{fncCoinTwoSamplePermutationTest}
\alias{fncCoinKruskalWallisTest}
\alias{fncCoinFriedmanTest}
\alias{fncCoinTwoWayTable}
\alias{fncCoinCochraneMHTest}
\alias{fncCoinAnsariBradleyTest}
\alias{fncCoinFlignerKilleenTest}
\alias{fncCoinSpearmanTest}
\alias{fncCoinMaxstatTest}
\alias{fncCoinMaxstatSurvTest}
\alias{fncCoinSurvTest}
\alias{enterTableMHTest}
\alias{fncCoinMHTest}
\alias{NumericOrDate}
\alias{startStop}
\description{Internal RcmdrcoinPlugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}
