\name{RcmdrcoinPlugin-package}
\alias{RcmdrcoinPlugin-package}
\alias{RcmdrcoinPlugin}
\docType{package}
\title{
Install the coin Rcmdr Plug-In
}
\description{
This package provides an Rcmdr \dQuote{plug-in} based on the coin package for an easier access to the coin package functions.
}
\details{
\tabular{ll}{
Package: \tab RcmdrcoinPlugin\cr
Type: \tab Package\cr
Version: \tab 1.0-20\cr
Date: \tab 2011-07-04\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Daniel-Corneliu Leucuta <danny.ldc@gmail.com>

Maintainer: Daniel-Corneliu Leucuta <danny.ldc@gmail.com>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}
